[NOT-FOR-UPSTREAM] Add changes to boot Ubuntu Jammy
This commit includes the local changes necessary to properly boot an
Ubuntu Jammy OS image to test systemd changes.

To boot the OS image, run: `sudo /path/to/mkosi -f qemu`.

The host system is running Ubuntu Jammy that was upgraded from Ubuntu
Focal.

This patch is not appropriate for upstream inclusion since this is only
applicable for my current local system environment. It is committed here
for convenience.

Signed-off-by: James Raphael Tiovalen <jamestiotio@gmail.com>